fix: revise login page
